# api/api.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "api.proto",
        "version": "1"
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/demo.service.v1.Demo/Ping": {
            "get": {
                "summary": "/demo.service.v1.Demo/Ping",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.google.protobuf.Empty"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        },
        "/demo.service.v1.Demo/SayHello": {
            "get": {
                "summary": "/demo.service.v1.Demo/SayHello",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.google.protobuf.Empty"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        },
        "/demo.service.v1.FileSystem/MediaIDGet": {
            "get": {
                "summary": "/demo.service.v1.FileSystem/MediaIDGet",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.MediaIDResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "fakeID",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "timestamp",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "demo.service.v1.FileSystem"
                ]
            }
        },
        "/demo.service.v1.FileSystem/NewsURLGet": {
            "get": {
                "summary": "/demo.service.v1.FileSystem/NewsURLGet",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.NewsURLGetResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "fakeID",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "timestamp",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "demo.service.v1.FileSystem"
                ]
            }
        },
        "/demo.service.v1.FileSystem/Ping": {
            "get": {
                "summary": "/demo.service.v1.FileSystem/Ping",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.google.protobuf.Empty"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "demo.service.v1.FileSystem"
                ]
            }
        },
        "/demo.service.v1.TBK/KeyConvert": {
            "get": {
                "summary": "/demo.service.v1.TBK/KeyConvert",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.KeyConvertResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "FromKey",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "UserID",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "tags": [
                    "demo.service.v1.TBK"
                ]
            }
        },
        "/demo.service.v1.TBK/Ping": {
            "get": {
                "summary": "/demo.service.v1.TBK/Ping",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.google.protobuf.Empty"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "demo.service.v1.TBK"
                ]
            }
        },
        "/demo.service.v1.TBK/WithDraw": {
            "get": {
                "summary": "/demo.service.v1.TBK/WithDraw",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.WithDrawResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "userID",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "tags": [
                    "demo.service.v1.TBK"
                ]
            }
        },
        "/demo.service.v1.Wechat/BalanceTemplateMsgSend": {
            "get": {
                "summary": "/demo.service.v1.Wechat/BalanceTemplateMsgSend",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.google.protobuf.Empty"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "userID",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "orderID",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "earningTime",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "salary",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "balance",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "tags": [
                    "demo.service.v1.Wechat"
                ]
            }
        },
        "/demo.service.v1.Wechat/MatchedTemplateMsgSend": {
            "get": {
                "summary": "/demo.service.v1.Wechat/MatchedTemplateMsgSend",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.google.protobuf.Empty"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "userID",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "orderID",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "paidTime",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "alipayTotalPrice",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "rebate",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "tags": [
                    "demo.service.v1.Wechat"
                ]
            }
        },
        "/kratos-demo/say_hello": {
            "get": {
                "summary": "/kratos-demo/say_hello",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.HelloResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        }
    },
    "definitions": {
        ".demo.service.v1.BalanceTemplateMsgSendReq": {
            "type": "object",
            "properties": {
                "userID": {
                    "type": "string"
                },
                "orderID": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "earningTime": {
                    "type": "string"
                },
                "salary": {
                    "type": "string"
                },
                "balance": {
                    "type": "string"
                }
            }
        },
        ".demo.service.v1.HelloReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            },
            "required": [
                "name"
            ]
        },
        ".demo.service.v1.HelloResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        ".demo.service.v1.KeyConvertReq": {
            "type": "object",
            "properties": {
                "FromKey": {
                    "type": "string"
                },
                "UserID": {
                    "type": "string"
                }
            }
        },
        ".demo.service.v1.KeyConvertResp": {
            "type": "object",
            "properties": {
                "ItemURL": {
                    "type": "string"
                },
                "Price": {
                    "type": "string"
                },
                "Rebate": {
                    "type": "string"
                },
                "Coupon": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "PicURL": {
                    "type": "string"
                }
            }
        },
        ".demo.service.v1.MatchedTemplateMsgSendReq": {
            "type": "object",
            "properties": {
                "userID": {
                    "type": "string"
                },
                "orderID": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "paidTime": {
                    "type": "string"
                },
                "alipayTotalPrice": {
                    "type": "string"
                },
                "rebate": {
                    "type": "string"
                }
            }
        },
        ".demo.service.v1.MediaIDReq": {
            "type": "object",
            "properties": {
                "fakeID": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        ".demo.service.v1.MediaIDResp": {
            "type": "object",
            "properties": {
                "MediaID": {
                    "type": "string"
                }
            }
        },
        ".demo.service.v1.NewsURLGetReq": {
            "type": "object",
            "properties": {
                "fakeID": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        ".demo.service.v1.NewsURLGetResp": {
            "type": "object",
            "properties": {
                "URL": {
                    "type": "string"
                }
            }
        },
        ".demo.service.v1.WithDrawReq": {
            "type": "object",
            "properties": {
                "userID": {
                    "type": "string"
                }
            }
        },
        ".demo.service.v1.WithDrawResp": {
            "type": "object",
            "properties": {
                "Rebate": {
                    "type": "string"
                },
                "OrderIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        ".google.protobuf.Empty": {
            "type": "object",
            "properties": {},
            "description": "A generic empty message that you can re-use to avoid defining duplicated\n empty messages in your APIs. A typical example is to use it as the request\n or the response type of an API method. For instance:\n\n     service Foo {\n       rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);\n     }\n\n The JSON representation for `Empty` is empty JSON object `{}`."
        }
    }
}